refactor(condo): DOMA-4476 Move getPaymentsSum to upper level
